# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Gerenciamento de Hospedagem",
        "description": "Documentação do Gerenciamento de Hospedagem",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/api/v1/dailies": {
            "get": {
                "tags": [
                    "Diária"
                ],
                "summary": "Listar os preços da diárias",
                "operationId": "cf97b558e9baf7568c2d33401f5354a8",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token JWT gerado durante a autenticação. Deve ser enviado no formato: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer jwt_token_aqui"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exibe uma lista paginada dos preços da diárias, com a possibilidade de carregar atributos relacionados e aplicar filtros."
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Token inválido ou ausente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diária"
                ],
                "summary": "Criar preço da diária",
                "operationId": "3a3bbbca93f09b1d8943381731bb38f1",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token JWT gerado durante a autenticação. Deve ser enviado no formato: Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer jwt_token_aqui"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para criação de um preço da diária",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "type_of_room_id"
                                ],
                                "properties": {
                                    "price": {
                                        "description": "O preço da diária",
                                        "type": "number",
                                        "example": "55.25"
                                    },
                                    "type_of_room_id": {
                                        "description": "Id do tipo do quarto",
                                        "type": "number",
                                        "example": "15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipos de quarto criado com sucesso."
                    },
                    "401": {
                        "description": "Token ausente ou inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Token inválido ou ausente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "500": {
                        "description": "Ocorreu um erro ao tentar recuperar os tipos de quarto.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "type": "string",
                                            "example": "Ocorreu um erro ao tentar recuperar os tipos de quarto.."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}